RemanofacturingORM_UploadingService

+++ uploader/src/api/files/FileController.ts
@@ -1,6 +1,6 @@
 import { NextFunction, Request, Response } from "express";
 import { FileMapper } from "../../database/mappers/FileMapper";
-import { FileRequestInfo } from "../../database/model/File";
+import { FileQueryInfo, FileRequestInfo } from "../../database/model/File";
 import FileService from "../../services/FileService";
 import { BadRequestException, NotFoundException } from "../errorHandler/Exceptions";
 
@@ -12,7 +12,7 @@ export default class FileController {
         this.fileService = new FileService();
         this.mapToDTO = new FileMapper().toDTO;
     }
-    
+
     async createNewFile(req: Request, res: Response, next: NextFunction) {
         if (!req.file) {
             return next(
@@ -21,14 +21,20 @@ export default class FileController {
                 )
             );
         }
-        const { originalname: name, size, mimetype: mimeType, path } = req.file;
-        const requestInfo = { name, size, mimeType, path };
+        const { originalname: name, size, mimetype: mimeType, id, chunkSize } = req.file;
+        const requestInfo = { name, size, mimeType, fsId: id, chunkSize };
         const newFile = await this.fileService.createNew(requestInfo);
         return res.status(202).json(this.mapToDTO(newFile));
     }
 
     async findFileById(req: Request, res: Response, next: NextFunction) {
         const { id } = req.params;
+        const objectIDLength = 24;
+        if (id.length != objectIDLength) {
+            throw new BadRequestException(
+                "Id passed in must be a string of 12 bytes or a string of 24 hex characters or an integer"
+            );
+        }
         const file = await this.fileService.findById(id);
         if (!file) {
             return next(new NotFoundException());
@@ -36,13 +42,24 @@ export default class FileController {
         return res.status(200).json(this.mapToDTO(file));
     }
 
+    async findBinaryFileById(req: Request, res: Response, next: NextFunction) {
+        const { id } = req.params;
+        const objectIDLength = 24;
+        if (id.length != objectIDLength) {
+            throw new BadRequestException(
+                "Id passed in must be a string of 12 bytes or a string of 24 hex characters or an integer"
+            );
+        }
+        const file = await this.fileService.findBinaryById(id);
+        if (!file) {
+            return next(new NotFoundException());
+        }
+        return res.status(200).json(file);
+    }
+
     async findFilesByQuery(req: Request, res: Response, next: NextFunction) {
-        const { name, size, mimeType } = req.query;
-        const requestInfo: FileRequestInfo = {
-            name: String(name),
-            size: Number(size),
-            mimeType: String(mimeType),
-        };
+        const { id, name, size, mimeType, chunkSize } = req.query;
+        const requestInfo: FileQueryInfo = { ...req.query };
         const files = await this.fileService.findManyByQuery(requestInfo);
         if (!files.length) {
             return next(new NotFoundException());
@@ -77,15 +94,22 @@ export default class FileController {
             ...file,
             name,
         });
-        return res.status(200).json(this.mapToDTO(updateFile));
+        return res.status(200).json(updateFile);
     }
 
     async deleteFile(req: Request, res: Response, next: NextFunction) {
         const { id } = req.params;
-        const deleted = await this.fileService.deleteById(id);
-        if (!deleted.affected) {
-            return next(new NotFoundException());
+        const objectIDLength = 24;
+        if (id.length != objectIDLength) {
+            throw new BadRequestException(
+                "Id passed in must be a string of 12 bytes or a string of 24 hex characters or an integer"
+            );
         }
-        return res.status(200).json(deleted);
+        try {
+            const deleted = await this.fileService.deleteById(id);
+            return res.status(200).json(deleted);
+        } catch {
+            throw new NotFoundException();
+        }
     }
 }
